+++ Documentation/Changelog.md
@@ -35,6 +35,7 @@
 
 For more details, please refer to:
 
+* [MariaDB MaxScale 2.5.27 Release Notes](Release-Notes/MaxScale-2.5.27-Release-Notes.md)
 * [MariaDB MaxScale 2.5.26 Release Notes](Release-Notes/MaxScale-2.5.26-Release-Notes.md)
 * [MariaDB MaxScale 2.5.25 Release Notes](Release-Notes/MaxScale-2.5.25-Release-Notes.md)
 * [MariaDB MaxScale 2.5.24 Release Notes](Release-Notes/MaxScale-2.5.24-Release-Notes.md)